Added P1203C

+++ .idea/fileTemplates/internal/Class.java
@@ -64,7 +64,7 @@ public class ${NAME} {
         public long[] nextLongArray(int n) {
             long[] arr = new long[n];
             for (int i = 0; i < n; i++)
-                arr[i] = nextInt();
+                arr[i] = nextLong();
             return arr;
         }
